requirements.txt
appdirs==1.4.4
asgiref==3.5.2
attrs==22.1.0
boto3==1.26.32
botocore==1.29.32
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-filter==22.1
django-nine==0.2.7
djangorestframework==3.14.0
drf-yasg==1.21.4
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
exceptiongroup==1.0.4
Faker==15.3.4
fs==2.4.16
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
model-bakery==1.9.0
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
